% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{accuracy}
\alias{accuracy}
\title{accuracy: Accuracy given predictions and ground truth.}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{Logical or numerical vector of binary predictions.}

\item{truth}{Logical or numerical vector of binary ground truth values.}
}
\value{
Numeric value - accuracy.
}
\description{
This function takes two vectors, pred and truth, as inputs, and returns the
accuracy
}
\examples{
accuracy(c(1,1,0,1,0), c(1,0,0,1,0))
}
