% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{adj_R2}
\alias{adj_R2}
\title{adj_R2: Adjusted R squared given predictions and ground truth.}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{Continuous numerical vector of predictions.}

\item{truth}{Numeric value that indicates the number of model parameters, excluding the intercept.}
}
\value{
Numeric value - Adjusted R Squared.
}
\description{
This function takes two vectors, pred and truth, as inputs, and returns the
adjusted R squared.
}
\examples{
adj_R2(c(2.6,2.2,4.2,3.9,6.0), c(2.3,3.2,4.7,3.9,3.3), n_p = 2)
}
