% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{classes}
\alias{classes}
\title{classes: Classes of variables in a tibble.}
\usage{
classes(x)
}
\arguments{
\item{x}{Tibble.}
}
\value{
A character vector with the identified classes.
}
\description{
This function takes a tibble x as an input, finds the class of each variable,
and returns a character vector with the corresponding classes.
}
\examples{
tib = dplyr::tibble(x = c("a", "b"), y = c(1, 2))
classes(tib)
}
