% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{df_scale}
\alias{df_scale}
\title{df_scale: Scaled numerical variables.}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{Tibble.}

\item{center}{Logical value that indicates if the columns should be centered.}

\item{scale}{Logical value that indicates if the columns should be scaled}
}
\value{
Scaled tibble.
}
\description{
This function takes a tibble x as an input, finds the class of each variable,
and returns a tibble in which the numeric variables have been scaled.
}
\examples{
tib = dplyr::tibble(x = c("a", "b", "c"), y = c(1, 2, 3), z = c(10, 2, 7))
df_scale(tib, center = TRUE, scale = TRUE)
}
