% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{f1}
\alias{f1}
\title{f1: F1-Score given predictions and ground truth.}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{Logical or numerical vector of binary predictions.}

\item{truth}{Logical or numerical vector of binary ground truth values}
}
\value{
Numeric value - F1-Score.
}
\description{
This function takes two vectors, pred and truth, as inputs, and returns the
F1-Score
}
\examples{
f1(c(1,1,0,1,0), c(1,0,0,1,0))
}
