% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{is_max}
\alias{is_max}
\title{is_max: Minimum value in atomic vector.}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{Atomic vector.}

\item{na.rm}{Logical value that indicates if NA values should be removed from x.}
}
\value{
A logical vector with TRUE where x equals its maximum value.
}
\description{
This function takes an atomic vector x as an input, finds the maximum value, and
returns a logical vector with TRUE where x equals its maximum value, and FALSE
for all other positions.
}
\examples{
av = c(1,2,3,4,5,6,7,9)
is_max(av, na.rm=TRUE)
}
