% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{log_likelihood_chisq: Log-likelihood chi-square distribution.}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{Numeric vector (sample).}

\item{df}{Degrees of freedom.}
}
\value{
Numeric value - Log-likelihood.
}
\description{
This function takes a vector x as an input, and returns the Log-likelihood
under the chi-squared distribution with the corresponding parametrizations.
}
\examples{
x = rchisq(100, df = 4)
log_likelihood_chisq(x = x, df = 4)
}
