% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{log_likelihood_f: Log-likelihood F distribution.}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{Numeric vector (sample).}

\item{df1}{Degrees of freedom.}

\item{df2}{Degrees of freedom.}
}
\value{
Numeric value - Log-likelihood.
}
\description{
This function takes a vector x as an input, and returns the Log-likelihood
under the F distribution with the corresponding parametrizations.
}
\examples{
x = rf(100, df1 = 2, df2 = 5)
log_likelihood_f(x = x, df1 = 2, df2 = 5)
}
