% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{log_likelihood_norm: Log-likelihood normal distribution.}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{Numeric vector (sample).}

\item{mean}{Mean of the distribution.}

\item{sd}{Standard deviation of the distribution.}
}
\value{
Numeric value - Log-likelihood.
}
\description{
This function takes a vector x as an input, and returns the Log-likelihood
under the normal distribution with the corresponding parametrizations.
}
\examples{
x = rnorm(100, mean = 0, sd = 1)
log_likelihood_norm(x = x, mean = 0, sd = 1)
}
