% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{log_likelihood_t: Log-likelihood t distribution.}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{Numeric vector (sample).}

\item{df}{Degrees of freedom.}
}
\value{
Numeric value - Log-likelihood.
}
\description{
This function takes a vector x as an input, and returns the Log-likelihood
under the t distribution with the corresponding parametrizations.
}
\examples{
x = rt(100, df = 2)
log_likelihood_t(x = x, df = 2)
}
