% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{log_likelihood_unif: Log-likelihood uniform distribution.}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{Numeric vector (sample).}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}
}
\value{
Numeric value - Log-likelihood.
}
\description{
This function takes a vector x as an input, and returns the Log-likelihood
under the uniform distribution with the corresponding parametrizations.
}
\examples{
x = runif(100, min = 0, max = 10)
log_likelihood_unif(x = x, min = 0, max = 10)
}
