% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{r2}
\alias{r2}
\title{r2: R squared given predictions and ground truth.}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{Continuous numerical vector of predictions.}

\item{truth}{Continuous numerical vector of ground truth values}
}
\value{
Numeric value - R Squared.
}
\description{
This function takes two vectors, pred and truth, as inputs, and returns the
R squared
}
\examples{
r2(c(2.6,2.2,4.2,3.9,6.0), c(2.3,3.2,4.7,3.9,3.3))
}
