% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{rando}
\alias{rando}
\title{Rando: Sample dataframes and vectors.}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{Either an atomic vector or dataframe-like object.}

\item{n}{Number of samples or rows to be obtained from the input ().}

\item{replace}{Logical value that indicates if sample is done with replacement.}
}
\value{
An atomic vector or dataframe-like object of size n.
}
\description{
This function checks if the input x is an atomic vector or a dataframe-like
object, and applies the appropriate operation to obtain a sample of size n.
}
\examples{
av = c(1,2,3,4,5,6,7,9)
rando(av, n = 5, replace = FALSE)

df = data.frame(x = c(1,2,3,4,5,6,7,8,9), y = c(9,8,7,6,5,4,3,2,1))
rando(df, n = 5, replace = TRUE)
}
