% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{rep_mat}
\alias{rep_mat}
\title{rep_mat: Replicates matrix in row and column dimension.}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{Matrix or dataframe.}

\item{M}{Integer that indicates the replicates of x in the row dimension}

\item{N}{Integer that indicates the replicates of x in the column dimension}
}
\value{
A matrix with replicates of x in the row and the column dimension.
}
\description{
This function takes a matrix or a dataframe x as an input, converts it to
a matrix, and returns a matrix with M replicates of x in the row dimension
an dN replicates of x in the column dimension.
}
\examples{
x = matrix(1:25,ncol=5)
rep_mat(x, M = 3, N = 2)
}
